+++ src/test/java/com/pivovarit/collectors/FunctionalTest.java
@@ -33,9 +33,9 @@ import java.util.stream.Stream;
 import static com.pivovarit.collectors.ParallelCollectors.parallel;
 import static com.pivovarit.collectors.ParallelCollectors.parallelToOrderedStream;
 import static com.pivovarit.collectors.ParallelCollectors.parallelToStream;
-import static com.pivovarit.collectors.infrastructure.TestUtils.incrementAndThrow;
-import static com.pivovarit.collectors.infrastructure.TestUtils.returnWithDelay;
-import static com.pivovarit.collectors.infrastructure.TestUtils.runWithExecutor;
+import static com.pivovarit.collectors.TestUtils.incrementAndThrow;
+import static com.pivovarit.collectors.TestUtils.returnWithDelay;
+import static com.pivovarit.collectors.TestUtils.runWithExecutor;
 import static java.lang.String.format;
 import static java.time.Duration.ofMillis;
 import static java.util.function.Function.identity;
